fix url structure
