+++ pages/conversations/[slug].tsx
@@ -5,17 +5,17 @@ import {
   useStoryblokState,
 } from '@storyblok/react';
 import { GetStaticPathsContext, GetStaticPropsContext, NextPage } from 'next';
-import NoDataAvailable from '../../../components/common/NoDataAvailable';
-import StoryblokResourceShortPage, {
-  StoryblokResourceShortPageProps,
-} from '../../../components/storyblok/StoryblokResourceShortPage';
-import { getStoryblokPageProps } from '../../../utils/getStoryblokPageProps';
+import NoDataAvailable from '../../components/common/NoDataAvailable';
+import StoryblokResourceConversationPage, {
+  StoryblokResourceConversationPageProps,
+} from '../../components/storyblok/StoryblokResourceConversationPage';
+import { getStoryblokPageProps } from '../../utils/getStoryblokPageProps';
 
 interface Props {
   story: ISbStoryData | null;
 }
 
-const ResourceShortOverview: NextPage<Props> = ({ story }) => {
+const ResourceConversationOverview: NextPage<Props> = ({ story }) => {
   story = useStoryblokState(story);
 
   if (!story) {
@@ -24,8 +24,8 @@ const ResourceShortOverview: NextPage<Props> = ({ story }) => {
 
   return (
     <>
-      <StoryblokResourceShortPage
-        {...(story.content as StoryblokResourceShortPageProps)}
+      <StoryblokResourceConversationPage
+        {...(story.content as StoryblokResourceConversationPageProps)}
         storyId={story.id}
       />
     </>
@@ -35,17 +35,18 @@ const ResourceShortOverview: NextPage<Props> = ({ story }) => {
 export async function getStaticProps({ locale, preview = false, params }: GetStaticPropsContext) {
   const slug = params?.slug instanceof Array ? params.slug.join('/') : params?.slug;
 
-  const storyblokProps = await getStoryblokPageProps(`shorts/${slug}`, locale, preview, {
-    resolve_relations: ['related_content', 'related_exercises'],
+  const storyblokProps = await getStoryblokPageProps(`conversations/${slug}`, locale, preview, {
+    resolve_relations: ['related_content.Course', 'related_content.Session'],
   });
 
+  console.log('storyblokProps', storyblokProps);
   return {
     props: {
       ...storyblokProps,
       messages: {
-        ...require(`../../../messages/shared/${locale}.json`),
-        ...require(`../../../messages/navigation/${locale}.json`),
-        // ...require(`../../../messages/resources/${locale}.json`),
+        ...require(`../../messages/shared/${locale}.json`),
+        ...require(`../../messages/navigation/${locale}.json`),
+        // ...require(`../../messages/resources/${locale}.json`),
       },
     },
     revalidate: 3600, // revalidate every hour
@@ -56,24 +57,24 @@ export async function getStaticPaths({ locales }: GetStaticPathsContext) {
   const isProduction = process.env.NEXT_PUBLIC_ENV === 'production';
   let sbParams: ISbStoriesParams = {
     version: isProduction ? 'published' : 'draft',
-    starts_with: 'shorts/',
+    starts_with: 'conversations/',
     filter_query: {
       component: {
-        in: 'resource_short',
+        in: 'resource_conversation',
       },
     },
   };
 
   const storyblokApi = getStoryblokApi();
-  let shorts = await storyblokApi.getAll('cdn/links', sbParams);
+  let conversations = await storyblokApi.getAll('cdn/links', sbParams);
 
   let paths: any = [];
 
-  shorts.forEach((short: Partial<ISbStoryData>) => {
-    if (!short.slug || (isProduction && !short.published)) return;
+  conversations.forEach((conversation: Partial<ISbStoryData>) => {
+    if (!conversation.slug || (isProduction && !conversation.published)) return;
 
     // get array for slug because of catch all
-    const slug = short.slug;
+    const slug = conversation.slug;
     let splittedSlug = slug.split('/');
 
     if (locales) {
@@ -90,4 +91,4 @@ export async function getStaticPaths({ locales }: GetStaticPathsContext) {
   };
 }
 
-export default ResourceShortOverview;
+export default ResourceConversationOverview;
